+++ gulpfile.js
@@ -12,7 +12,6 @@ gulp.task("vendor", function () {
     "./node_modules/bootstrap/dist/js/**/bootstrap.min.js*",
     "./node_modules/@popperjs/core/dist/umd/**/*.min.js",
     "./node_modules/highlight.js/styles/**/*.min.css",
-    "./node_modules/masonry-layout/dist/**/*.min.js",
     "./node_modules/@fixhq/sweetalert2/dist/**/*.min.*",
   ];