+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **[Edit]** Correcting `sweetalert2` includes
 - **[Deps]** Remove `fitvids.js` replacing with CSS
+- **[Deps]** Remove `masonry-layout` replacing with CSS
 - **[Deps]** Remove `jquery.backstretch` as it is unused
 
 ## 2025.09.13 / v0.8.0